requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
exceptiongroup==1.1.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
lxml==4.9.3
numpy==1.25.2
outcome==1.2.0
pandas==2.1.0
psycopg2==2.9.7
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.20
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0

fastapi~=0.103.2
APScheduler~=3.10.4
uvicorn~=0.23.2